Better factorial implementation
