+++ Common/src/main/java/at/petrak/hexcasting/common/casting/arithmetic/operator/OperatorUtils.kt
@@ -5,8 +5,12 @@ import at.petrak.hexcasting.api.casting.iota.DoubleIota
 import at.petrak.hexcasting.api.casting.iota.Iota
 import at.petrak.hexcasting.api.casting.iota.ListIota
 import at.petrak.hexcasting.api.casting.mishaps.MishapInvalidIota
+import kotlin.math.PI
 import kotlin.math.abs
+import kotlin.math.exp
+import kotlin.math.ln
 import kotlin.math.roundToInt
+import kotlin.math.sqrt
 
 fun Iterator<IndexedValue<Iota>>.nextList(argc: Int = 0): SpellList {
     val (idx, x) = this.next()
@@ -65,3 +69,26 @@ fun Iterator<IndexedValue<Iota>>.nextPositiveIntUnderInclusive(max: Int, argc: I
  */
 fun Double.asDoubleBetween(min: Double, max: Double, idx: Int) = if (this in min .. max) this
     else throw MishapInvalidIota.of(DoubleIota(this), idx, "double.between", min, max)
+
+/**
+ * Takes the standard factorial if `arg` is a positive integer, otherwise uses the gamma function.
+ */
+fun generalFactorial(arg: Double): Double {
+    val argInt = arg.roundToInt()
+    if (arg >= 0 && DoubleIota.tolerates(arg, argInt.toDouble())) {
+        // normal integer factorial
+        var result: Long = 1
+        for (factor in 2..argInt) {
+            result *= factor
+        }
+        return result.toDouble()
+    }
+    // gamma function of arg+1 for non-integer factorial
+    // https://introcs.cs.princeton.edu/java/91float/Gamma.java.html
+    val x = arg + 1
+    val ser = ( 1.0 + 76.18009173    / (x + 0)   - 86.50532033    / (x + 1)
+                    + 24.01409822    / (x + 2)   -  1.231739516   / (x + 3)
+                    +  0.00120858003 / (x + 4)   -  0.00000536382 / (x + 5) )
+    val logGamma = (x - 0.5) * ln(x + 4.5) - (x + 4.5) + ln(ser * sqrt(2*PI))
+    return exp(logGamma)
+}
